Organize worker modules
